requirements.txt
# Core Web Framework
Flask==3.1.1
Flask-CORS==4.0.0
Werkzeug==3.1.3
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2

# Local LLM Integration (PRIMARY)
ollama==0.1.7
llama-cpp-python==0.2.27

# Optional Cloud APIs (FALLBACK ONLY)
# openai==1.12.0  # Uncomment only if user wants cloud fallback
# anthropic==0.18.1  # Uncomment only if user wants cloud fallback

# Lightweight Local AI
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# Code Analysis and Processing
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.3
tree-sitter-typescript==0.20.3
ast-decompiler==0.7.0

# File System Monitoring
watchdog==3.0.0

# Code Formatting and Linting
black==25.1.0
ruff==0.12.2
isort==5.13.2
mypy==1.8.0

# Utilities
click==8.2.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
mypy_extensions==1.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.1
